pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "presonus-io24"
version = "0.1.0"
description = "CLI control utility for Presonus Revelator IO24 audio interface"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Developer", email = "dev@example.com"}
]
keywords = [
    "presonus",
    "audio",
    "usb",
    "cli",
    "linux",
    "revelator",
    "io24",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: System :: Hardware :: Hardware Drivers",
    "Topic :: Utilities",
]
dependencies = [
    "pyusb>=1.2.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
presonus-io24 = "cli.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["presonus*", "cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
